+++ packages/spacecat-shared-data-access/src/models/site.js
@@ -41,6 +41,7 @@ const Site = (data = {}) => {
   self.getGitHubURL = () => self.state.gitHubURL;
   self.getOrganizationId = () => self.state.organizationId;
   self.isLive = () => self.state.isLive;
+  self.getIsLiveToggledAt = () => self.state.isLiveToggledAt;
 
   // TODO: updating the baseURL is not supported yet, it will require a transact write
   //  on dynamodb (put then delete) since baseURL is part of the primary key, something like:
@@ -141,6 +142,7 @@ const Site = (data = {}) => {
    */
   self.toggleLive = () => {
     self.state.isLive = !self.state.isLive;
+    self.state.isLiveToggledAt = new Date().toISOString();
     return self;
   };