+++ packages/spacecat-shared-data-access/test/it/generateSampleData.js
@@ -194,6 +194,7 @@ export default async function generateSampleData(
       gitHubURL: `https://github.com/org-${i}/test-repo`,
       organizationId: organizations[i % 3].id,
       isLive: true,
+      isLiveToggledAt: nowIso,
       GSI1PK: config.pkAllSites,
       createdAt: nowIso,
       updatedAt: nowIso,